// src/interfaces/satelliteCampus.d.ts
export interface I_Satellite_Campus {
    _id: string,
    campusName: string,
    mainCampusId: string,
    schoolName: string,
    email: string,
    mobileNo: string,
    currency: string,
    currencySymbol: string,
    city: string,
    state: string,
    address: string,
    createdAt: string,
    updatedAt: date | undefined,
    deletedAt: date | undefined
}